Light and Collision Total Rework Begin
Game is stable and functional, but not fully reworked.

Creating a global collision manager that will handle all collision logic to allow more precise control and hopefully get around issues with overlapping collisions.

Anything capable of dealing damage need to be a child of the hurtbox parent, or spawn a separate hurtbox object that is.

This means spells that glow (like the firebolt) now instead spawn a separate light object