Massive Collision Rework
Collisions are now centralized in a collision manager. Anything that can deal damage (to or from player) either needs to be a child of the hurtbox parent or create a separate hurtbox object that is.

Each must have a damage value, an on hit funciton, a team variable (enemy or player), and an active bool